requirements.txt
torch
ipykernel
numpy
pandas
scikit-learn
matplotlib
torchvision 
cupy-cuda12x
numpy